requirements.txt
Django~=5.1.3
djangorestframework~=3.15.2
python-dotenv~=1.0.1
psycopg2>=2.8
astropy>=6.0